setup beginnings of thread separation
