+++ .gitignore
@@ -12,3 +12,6 @@ cmake-build-debug
 cmake-build-release
 build
 *.iso
+debug
+release
+.vs
