+++ src/core/gs.hpp
@@ -1,56 +1,10 @@
 #ifndef GS_HPP
 #define GS_HPP
 #include <cstdint>
+#include <thread>
 #include "gscontext.hpp"
+#include "circularFIFO.h"
 
-struct PRMODE
-{
-    bool gourand_shading;
-    bool texture_mapping;
-    bool fog;
-    bool alpha_blend;
-    bool antialiasing;
-    bool use_UV;
-    bool use_context2;
-    bool fix_fragment_value;
-};
-
-struct PRIM_REG
-{
-    uint8_t prim_type;
-    bool gourand_shading;
-    bool texture_mapping;
-    bool fog;
-    bool alpha_blend;
-    bool antialiasing;
-    bool use_UV;
-    bool use_context2;
-    bool fix_fragment_value;
-};
-
-struct RGBAQ_REG
-{
-    uint8_t r, g, b, a;
-    float q;
-};
-
-struct BITBLTBUF_REG
-{
-    uint32_t source_base;
-    uint32_t source_width;
-    uint8_t source_format;
-    uint32_t dest_base;
-    uint32_t dest_width;
-    uint8_t dest_format;
-};
-
-struct TRXPOS_REG
-{
-    uint16_t source_x, source_y;
-    uint16_t dest_x, dest_y;
-    uint16_t int_source_x, int_dest_x;
-    uint8_t trans_order;
-};
 
 //reg reg
 struct TRXREG_REG
@@ -96,8 +50,7 @@ struct TEXCLUT_REG
 {
     uint16_t width, x, y;
 };
-
-struct GS_IMR
+struct GS_IMR //Interrupt masking
 {
     bool signal;
     bool finish;
@@ -105,22 +58,59 @@ struct GS_IMR
     bool vsync;
     bool rawt; //Rectangular Area Write Termination
 };
-
-struct Vertex
+struct GS_CSR//sets mode on write, obtains status on read
 {
-    int32_t x, y, z;
-
-    RGBAQ_REG rgbaq;
-    UV_REG uv;
-    float s, t;
-    void to_relative(XYOFFSET xyoffset)
-    {
-        x -= xyoffset.x;
-        y -= xyoffset.y;
-    }
+	bool signal;
+	bool finish;
+	bool hsint, vsint;
+	bool edwint; //Rectangular Area Write Termination Interrupt Control
+	bool flush;
+	bool reset;
+	bool nfield;
+	bool field; //"field displayed currently" AKA is_odd_frame
+	bool fifo[2];//host interface FIFO status
+	uint8_t rev;//revision No. of the GS?
+	uint8_t id;//ID of the GS?
+	//note: there are padding bits not included in this struct
 };
 
 class INTC;
+
+enum GS_command:uint8_t { write64, write64_privileged, write32_privileged,
+    set_rgba, set_stq, set_uv, set_xyz, set_q };
+union GS_message_payload {
+    struct {
+        uint32_t addr;
+        uint64_t value;
+    } write64_payload;
+    struct {
+        uint32_t addr;
+        uint32_t value;
+    } write32_payload;
+    struct {
+        uint8_t r, g, b, a;
+    } rgba_payload;
+    struct {
+        uint32_t s, t, q;
+    } stq_payload;
+    struct {
+        uint16_t u, v;
+    } uv_payload;
+    struct {
+        uint32_t x, y, z;
+        bool drawing_kick;
+    } xyz_payload;
+    struct {
+        float q;
+    } q_payload;
+    struct {} no_payload;
+};
+struct GS_message {
+    GS_command type;
+    GS_message_payload payload;
+};
+typedef CircularFifo<GS_message, 10000> gs_fifo;
+
 
 class GraphicsSynthesizer
 {
@@ -132,75 +122,23 @@ class GraphicsSynthesizer
         uint8_t* local_mem;
         uint8_t CRT_mode;
 
-        //CSR/IMR stuff - to be merged into structs
-        bool VBLANK_generated;
-        bool VBLANK_enabled;
-        bool is_odd_frame;
-        bool FINISH_enabled;
-        bool FINISH_generated;
-        bool FINISH_requested;
-
-        GS_IMR IMR;
-
-        GSContext context1, context2;
-        GSContext* current_ctx;
-
-        PRIM_REG PRIM;
-        RGBAQ_REG RGBAQ;
-        UV_REG UV;
-        ST_REG ST;
-        TEXCLUT_REG TEXCLUT;
-        bool DTHE;
-        bool COLCLAMP;
-        bool use_PRIM;
-
-        BITBLTBUF_REG BITBLTBUF;
-        TRXPOS_REG TRXPOS;
-        TRXREG_REG TRXREG;
-        uint8_t TRXDIR;
-        uint8_t BUSDIR;
-        int pixels_transferred;
 
         //Used for unpacking PSMCT24
         uint32_t PSMCT24_color;
         int PSMCT24_unpacked_count;
 
+        //Privileged registers
         PMODE_REG PMODE;
         SMODE SMODE2;
         DISPFB DISPFB1, DISPFB2;
         DISPLAY DISPLAY1, DISPLAY2;
 
-        Vertex current_vtx;
-        Vertex vtx_queue[3];
-        unsigned int num_vertices;
-
-        static const unsigned int max_vertices[8];
-
-        uint32_t get_word(uint32_t addr);
-        void set_word(uint32_t addr, uint32_t value);
-
-        uint32_t read_PSMCT32_block(uint32_t base, uint32_t width, uint32_t x, uint32_t y);
-        uint16_t read_PSMCT16_block(uint32_t base, uint32_t width, uint32_t x, uint32_t y);
-        void write_PSMCT32_block(uint32_t base, uint32_t width, uint32_t x, uint32_t y, uint32_t value);
-        void write_PSMCT16_block(uint32_t base, uint32_t width, uint32_t x, uint32_t y, uint16_t value);
-
-        bool depth_test(int32_t x, int32_t y, uint32_t z);
-
-        void tex_lookup(int16_t u, int16_t v, const RGBAQ_REG& vtx_color, RGBAQ_REG& tex_color);
-        void clut_lookup(uint8_t entry, RGBAQ_REG& tex_color, bool eight_bit);
-        void clut_CSM2_lookup(uint8_t entry, RGBAQ_REG& tex_color);
-        void vertex_kick(bool drawing_kick);
-        void draw_pixel(int32_t x, int32_t y, uint32_t z, RGBAQ_REG& color, bool alpha_blending);
-        void render_primitive();
-        void render_point();
-        void render_line();
-        void render_triangle();
-        void render_sprite();
-        void write_HWREG(uint64_t data);
-        void unpack_PSMCT24(uint64_t data, int offset);
-        void host_to_host();
-
-        int32_t orient2D(const Vertex &v1, const Vertex &v2, const Vertex &v3);
+        GS_CSR CSR;
+        GS_IMR IMR;
+        uint8_t BUSDIR;
+        gs_fifo MessageQueue; //ring buffer size
+        //EXTBUF, EXTDATA, EXTWRITE, BGCOLOR, SIGLBLID not currently implemented
+		
     public:
         GraphicsSynthesizer(INTC* intc);
         ~GraphicsSynthesizer();
@@ -219,8 +157,8 @@ class GraphicsSynthesizer
 
         void set_CRT(bool interlaced, int mode, bool frame_mode);
 
-        uint32_t read32_privileged(uint32_t addr);
-        uint64_t read64_privileged(uint32_t addr);
+        uint32_t read32_privileged(uint32_t addr); //fails
+        uint64_t read64_privileged(uint32_t addr); //fails
         void write32_privileged(uint32_t addr, uint32_t value);
         void write64_privileged(uint32_t addr, uint64_t value);
         void write64(uint32_t addr, uint64_t value);
@@ -231,15 +169,5 @@ class GraphicsSynthesizer
         void set_Q(float q);
         void set_XYZ(uint32_t x, uint32_t y, uint32_t z, bool drawing_kick);
 };
-
-inline uint32_t GraphicsSynthesizer::get_word(uint32_t addr)
-{
-    return *(uint32_t*)&local_mem[addr];
-}
-
-inline void GraphicsSynthesizer::set_word(uint32_t addr, uint32_t value)
-{
-    *(uint32_t*)&local_mem[addr] = value;
-}
 
 #endif // GS_HPP
